documented OrderManager setup in host app... all complex API-driven components will follow this pattern
